requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
chromadb>=0.4.22
beautifulsoup4>=4.12.2
requests>=2.31.0
lxml>=4.9.3
matplotlib>=3.8.2
seaborn>=0.13.0
pandas>=2.1.4
numpy>=1.26.2
python-dotenv>=1.0.0
pydantic>=2.5.3
typing-extensions>=4.9.0
flask>=3.0.0
gunicorn>=21.2.0
flask-cors>=4.0.0

